trading_bot/requirements.txt
python-binance
click
python-dotenv
rich
